requirements.txt
PyYaml==3.10
selenium==2.37.2